skills/fusionauth-webhooks/examples/fastapi/requirements.txt
fastapi>=0.128.0
uvicorn>=0.34.0
PyJWT>=2.11.0
python-dotenv>=1.0.0
pytest>=9.0.2
httpx>=0.28.1
